+++ database/migrations/2026_04_24_062213_create_predictions_table.php
@@ -13,13 +13,13 @@ return new class extends Migration
     {
         Schema::create('predictions', function (Blueprint $table) {
             $table->id('predictions_id');
-            $table->integer('enrollment_batch_id');
+            $table->unsignedBigInteger('enrollment_batch_id');
             $table->foreign('enrollment_batch_id')->references('enrollment_batch_id')->on('enrollment_batches');
             $table->decimal('predicted_total');
             $table->decimal('predicted_male');
             $table->decimal('predicted_female');
             $table->decimal('confidence');
-            $table->integer('mlmodel_id');
+            $table->unsignedBigInteger('mlmodel_id');
             $table->foreign('mlmodel_id')->references('mlmodel_id')->on('mlmodels');
             $table->timestamps();
         });
